refactor(core): Acquire the expression isolate inside getNodesWithUnregisteredWebhooks (no-changelog)
Follow-up to #33341. That fix acquired the isolate in the activator, bracketing its call to the webhook trigger registrar. Move the acquire down into WebhookTriggerRegistrar.getNodesWithUnregisteredWebhooks — the method that actually resolves each trigger's path/httpMethod expressions via getWebhookTriggers — so the acquire lives with the code that needs it rather than one layer up in the caller.

Behavior is unchanged; the acquire/release ordering and release-on-throw tests move down to the registrar test suite alongside the code.

Related: CAT-3636
