+++ packages/cli/src/workflows/triggers/__tests__/workflow-trigger-activator.test.ts
@@ -167,54 +167,6 @@ describe('WorkflowTriggerActivator', () => {
 			);
 		});
 
-		test('brackets the registrar call with an acquired isolate so path expressions resolve', async () => {
-			const callOrder: string[] = [];
-			vi.spyOn(WorkflowExpression.prototype, 'acquireIsolate').mockImplementation(async () => {
-				callOrder.push('acquire');
-			});
-			vi.spyOn(WorkflowExpression.prototype, 'releaseIsolate').mockImplementation(async () => {
-				callOrder.push('release');
-			});
-			vi.spyOn(WorkflowExecuteAdditionalData, 'getBase').mockResolvedValue(
-				mock<IWorkflowExecuteAdditionalData>(),
-			);
-			const webhookTriggerRegistrar = mock<WebhookTriggerRegistrar>();
-			webhookTriggerRegistrar.getNodesWithUnregisteredWebhooks.mockImplementation(async () => {
-				callOrder.push('resolve');
-				return new Set(['w']);
-			});
-			const activator = buildActivator({ webhookTriggerRegistrar });
-
-			await activator.getNodesWithUnregisteredWebhooks(
-				mock<WorkflowEntity>({ id: 'wf-1', name: 'Test workflow', staticData: {}, settings: {} }),
-				{ nodes: [node('w', 'webhook')], connections: {} },
-			);
-
-			expect(callOrder).toEqual(['acquire', 'resolve', 'release']);
-		});
-
-		test('releases the isolate when the registrar throws', async () => {
-			vi.spyOn(WorkflowExecuteAdditionalData, 'getBase').mockResolvedValue(
-				mock<IWorkflowExecuteAdditionalData>(),
-			);
-			const releaseIsolate = vi
-				.spyOn(WorkflowExpression.prototype, 'releaseIsolate')
-				.mockResolvedValue(undefined);
-			vi.spyOn(WorkflowExpression.prototype, 'acquireIsolate').mockResolvedValue(undefined);
-			const webhookTriggerRegistrar = mock<WebhookTriggerRegistrar>();
-			webhookTriggerRegistrar.getNodesWithUnregisteredWebhooks.mockRejectedValue(new Error('boom'));
-			const activator = buildActivator({ webhookTriggerRegistrar });
-
-			await expect(
-				activator.getNodesWithUnregisteredWebhooks(
-					mock<WorkflowEntity>({ id: 'wf-1', name: 'Test workflow', staticData: {}, settings: {} }),
-					{ nodes: [node('w', 'webhook')], connections: {} },
-				),
-			).rejects.toThrow('boom');
-
-			expect(releaseIsolate).toHaveBeenCalledTimes(1);
-		});
-
 		test('returns empty without calling the registrar when there are no trigger nodes', async () => {
 			const webhookTriggerRegistrar = mock<WebhookTriggerRegistrar>();
 			const activator = buildActivator({ webhookTriggerRegistrar });
