+++ packages/cli/src/workflows/triggers/workflow-trigger-activator.ts
@@ -155,18 +155,11 @@ export class WorkflowTriggerActivator {
 			workflowSettings: dbWorkflow.settings,
 		});
 
-		// Resolving webhook triggers evaluates each node's `path`/`httpMethod`
-		// expressions (e.g. the Form Trigger's dynamic path), which needs an isolate.
-		await workflow.expression.acquireIsolate();
-		try {
-			return await this.webhookTriggerRegistrar.getNodesWithUnregisteredWebhooks(
-				workflow,
-				additionalData,
-				desiredNodes,
-			);
-		} finally {
-			await workflow.expression.releaseIsolate();
-		}
+		return await this.webhookTriggerRegistrar.getNodesWithUnregisteredWebhooks(
+			workflow,
+			additionalData,
+			desiredNodes,
+		);
 	}
 
 	/**
